requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0

# Hugging Face
transformers>=4.40.0
sentence-transformers>=3.0.0
huggingface-hub>=0.20.0
accelerate>=0.30.0
bitsandbytes>=0.41.0  # Para quantização 4-bit (QLoRA)

# Vector Store - ChromaDB
chromadb>=0.5.0

# Document Processing
pypdf>=5.0.0
pdfplumber>=0.11.0

# Text Processing
tiktoken>=0.7.0
# unstructured - removido temporariamente (problemas com Python 3.13)

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
numpy>=1.24.0  # NumPy 1.x é mais estável, mas 2.x também funciona
pandas>=2.0.0

# Optional Tools (Wikipedia, calculadora, etc)
wikipedia>=1.4.0
numexpr>=2.10.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Interface
gradio>=4.0.0

# Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Progress bars
tqdm>=4.65.0

# Torch (instala automaticamente a versão correta para o sistema)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Monitoring (opcional)
# langsmith - opcional, pode instalar depois se necessário